feat(loans): support half-month loan terms (e.g. 5.5)
Contract-of-Service contracts can run a half-month, but a 5.5 term was
dropped in two places: parseTerms cast every value through (int), turning
5.5 into 5 and then deduping it away, and the term_months column was an
integer so it could not store the fraction anyway. Separately, the remaining
contract length was rounded to whole months, so a member with 5.28 months
left saw "5" and a 5.5 term was filtered out as exceeding the contract.

Changes:
- term_months migrates int -> decimal(5,2)
- parseTerms keeps fractional values (5.0 still displays as "5", 5.5 stays)
- contract length and remaining months round to the nearest half-month, so
  the SC max loan and the term filter both reflect 5.5
- amortization accepts a fractional term; the schedule runs ceil(term)
  periods with a partial final installment that absorbs rounding
- validation on every entry point (apply, OTP confirm, PIN confirm,
  eligibility) accepts numeric min 0.5 instead of integer
- the loan chatbot parses "5.5 months" and matches it against config terms

Verified: a 5.5-month ₱50,000 loan stores term 5.50, generates 6 schedule
rows where the last is exactly half, and the installments sum precisely to
the total payable.

+++ app/Http/Controllers/Api/ReportController.php
@@ -481,7 +481,7 @@ class ReportController extends Controller
     private function formatLedgerLoan($loan): array
     {
         $principal = (float) $loan->amount;
-        $interest  = round($principal * ((float) $loan->interest_rate / 100) * (int) $loan->term_months, 2);
+        $interest  = round($principal * ((float) $loan->interest_rate / 100) * (float) $loan->term_months, 2);
         $total     = $loan->total_payable; // principal + interest (exact)
 
         $running = $total;
